Try to clean up Weapon_Sword_Fire. Suppress warning in clang about system call in SVCmd_DeleteCharacter_f.

+++ g_sword.c
@@ -184,8 +184,8 @@ void sword_attack (edict_t *ent, vec3_t g_offset, int damage)
 
 void Weapon_Sword_Fire (edict_t *ent)
 {
-	int sword_bonus = 1;
-	int damage;
+	float sword_bonus = 1;
+	float damage;
 	float temp;
 	
 	// special rules; flag carrier can't use weapons
@@ -200,7 +200,7 @@ void Weapon_Sword_Fire (edict_t *ent)
 	temp = 0.8 + 0.007 * ent->myskills.weapons[WEAPON_SWORD].mods[1].current_level;
 
 	 if ((temp < 1.0) && (ent->client->ps.gunframe - 5 > 0))
-		damage *= pow(temp, ent->client->ps.gunframe - 5);
+		damage *= powf(temp, ent->client->ps.gunframe - 5);
 	
 	 //gi.dprintf("damage=%d\n", damage);
 
@@ -208,7 +208,7 @@ void Weapon_Sword_Fire (edict_t *ent)
 		gi.sound (ent, CHAN_WEAPON, gi.soundindex("misc/power1.wav") , 1, ATTN_NORM, 0);
 
      if ( ent->client->buttons & BUTTON_ATTACK )
-		sword_attack (ent, vec3_origin, damage);
+		sword_attack (ent, vec3_origin, (int)damage);
      ent->client->ps.gunframe++;
 }
